pyproject.toml
[project]
name = "webgenie-task-generator"
version = "0.1.0"
description = "WebGenie Task Generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [    
    "ansible-vault==2.1.0",
    "async-substrate-interface==1.0.0",
    "beautifulsoup4==4.12.3",
    "bert-score==0.3.13",
    "bt-decode==0.5.0a2",
    "bittensor==9.0.0",
    "bittensor-cli==9.0.0",
    "clip",
    "datasets==3.2.0",
    "ddt==1.6.0",
    "duckduckgo_search",
    "datasets",
    "einops",
    "fastapi",
    "lxml==5.3.0",
    "matplotlib-inline==0.1.7",
    "nltk",
    "openai",
    "peft",
    "pip-chill==1.0.3",
    "playwright==1.49.1",
    "pydantic==2.6",
    "python-dotenv==1.0.1",
    "scikit-learn==1.6.0",
    "shtab==1.6.5",
    "sqlalchemy",
    "tinycss2==1.4.0",
    "wandb==0.19.0",
    "numpy>=2.0.2",
    "colormath>=3.0.0",
    "scikit-image>=0.25.0",
    "sentence-transformers",
    "tensorflow>=2.18.0",
    "tf-keras",
    "uvicorn",
]
